fixes, not yet working

+++ packages/db/entities/ChangeRecord.ts
@@ -1,8 +1,8 @@
 import { ISpec } from '@ag/util'
-import { Column, Entity, PrimaryColumn } from 'typeorm'
-import { AppTable } from './appEntities'
-import { DbChange, DbEntityEdit } from './DbChange'
-import { DbEntity } from './DbEntity'
+import { Column, Connection, Entity, PrimaryColumn } from 'typeorm'
+import { simplifyEntity } from '../export'
+import { AppTable, appTable } from './appEntities'
+import { DbChange } from './DbChange'
 
 type ChangeType = 'add' | 'edit' | 'delete'
 
@@ -21,18 +21,36 @@ export class ChangeRecord<T extends {} = {}> {
     }
   }
 
-  static fromDbChange = (change: DbChange[]): ChangeRecord[] => {
-    const records: ChangeRecord[] = change.flatMap(({ table, t, adds, edits, deletes }) => [
-      ...(adds
-        ? adds.map(add => new ChangeRecord({ table, id: add.id, t, type: 'add', value: add }))
-        : []),
-      ...(edits
-        ? edits.map(edit => new ChangeRecord({ table, id: edit.id, t, type: 'edit', edit: edit.q }))
-        : []),
-      ...(deletes
-        ? deletes.map(id => new ChangeRecord({ table, id, t, type: 'delete' })) //
-        : []),
-    ])
+  static fromDbChange = (connection: Connection, change: DbChange[]): ChangeRecord[] => {
+    const records: ChangeRecord[] = change.flatMap(({ table, t, adds, edits, deletes }) => {
+      const entityMetadata = connection.entityMetadatas.find(emd => emd.target === appTable[table])
+      if (!entityMetadata) {
+        throw new Error('no entityMetadata for table ' + table)
+      }
+
+      return [
+        ...(adds
+          ? adds.map(
+              add =>
+                new ChangeRecord({
+                  table,
+                  id: add.id,
+                  t,
+                  type: 'add',
+                  value: simplifyEntity(add, entityMetadata),
+                })
+            )
+          : []),
+        ...(edits
+          ? edits.map(
+              edit => new ChangeRecord({ table, id: edit.id, t, type: 'edit', edit: edit.q })
+            )
+          : []),
+        ...(deletes
+          ? deletes.map(id => new ChangeRecord({ table, id, t, type: 'delete' })) //
+          : []),
+      ]
+    })
     return records
   }
 }
